Clean up some example CQL warnings (#3174)

+++ libraries/FatFSUSB/examples/Listfiles-USB/Listfiles-USB.ino
@@ -78,7 +78,7 @@ void printDirectory(String dirName, int numTabs) {
       // no more files
       break;
     }
-    for (uint8_t i = 0; i < numTabs; i++) {
+    for (int i = 0; i < numTabs; i++) {
       Serial.print('\t');
     }
     Serial.print(dir.fileName());
